refactor(node): extract cmd_init to commands/init.rs
One-shot bootstrap handler moves out of main.rs alongside the other
cmd_* extractions. Wraps Genesis::from_path / Genesis::mainnet +
Blockchain::new_with_genesis + a first save. Pure CLI; no consensus.

main.rs drops to 3675 lines (peak was 4780; -23%). cmd_start (~2700
LOC) is the only handler left inline — it owns the BFT engine,
libp2p swarm, and validator key load, which the orchestration layer
needs to drive directly.
